% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscacod.r
\name{buscacod}
\alias{buscacod}
\title{Función de búsqueda del código del grupo y la familia}
\usage{
buscacod(nomb)
}
\arguments{
\item{nomb}{Nombre científico de la especie o fragmento del nombre entre ""}
}
\description{
Busca el código del grupo, especie y la familia a partir del nombre o fragmento de la especie. Es independiente de mayúsculas y minúsculas y puede sacar más de una especie si una parte de sus nombres coincide.
}
\examples{
buscacod("sph")
}
\seealso{
Other datos_especies: \code{\link{buscaesp}},
  \code{\link{talpes.camp}}
}
\concept{datos_especies}
