% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan2.camp.r
\name{lan2.camp}
\alias{lan2.camp}
\title{Lances especiales de una campaña}
\usage{
lan2.camp(camp, dns)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}
}
\description{
Busca y devuelve los lances especiales de una campaña
}
\examples{
lan2.camp("P01","Porc")
}
