% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFauna.camp.r
\name{ListFauna.camp}
\alias{ListFauna.camp}
\title{Capturas medias por lance de cada especie de gr (grupo) en una campaña}
\usage{
ListFauna.camp(gr = "1", camp, dns, cor.time = TRUE, excl.sect = NA,
  incl2 = FALSE, kg = TRUE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros.}

\item{camp}{Campaña de la que se extrae el listado de especies capturadas: un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa", Medits "Medi"}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos, NA no excluye ninguno}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\value{
Devuelve un data.frame con las capturas medias por lance de cada especie capturada del grupo gr. Columnas: gr,esp,especie,peso(kg),número,nlan (nº de lances en que ha aparecido esp) Los valores NaN en las abundancias corresponden a especies que sólo han aparecido en los lances especiales, y que no puede calcularse la abundancia estratificada al no contar con áreas para los estratos en que aparecen
}
\description{
Muestra un listado de la biomasa y número medio de todas las especies capturadas del grupo gr
}
\examples{
ListFauna.camp(gr=1,camp="N12",dns="Cant",excl.sect=FALSE,incl2=FALSE)
}
