% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p95tal.camp.r
\name{p95tal.camp}
\alias{p95tal.camp}
\title{Da el percentil .95 de la distribucion de tallas estratificada}
\usage{
p95tal.camp(gr = 1, esp, camps, dns = "Cant", excl.sect = NA,
  perc = 0.95, cex.leg = 1.1)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos (4 equinodermos 5 invertebrados habitualmente no medidos)}

\item{esp}{ha de ser código de la especie seleccionada}

\item{camps}{Campañas de las que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Excluye el los sectores y estratos en cuestion, si NA usa toda el area.}

\item{perc}{El percentil de la distribucion de tallas a calcular, .95 es el valor utilizado habitualmente como indicador del maximo de la distribucione de tallas}
}
\value{
Devuelve un data.frame con columnas camp y p95 tal con el Percentil 95 de la correspondiente campaña, una con el peso o número de cada especie del grupo solicitado, lat, long, prof
}
\description{
Funcion auxiliar indicador de la MSFD: Da el percentil perc (por defecto 95, en decimal) de la distribución de tallas de la especie en las campañas seleccionadas.
}
\examples{
p95tal.camp(1,50,Nsh[7:27],"Cant")
}
