% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacagrid.r
\name{sacagrid}
\alias{sacagrid}
\title{Saca el grid de cuadrículas para Porcupine}
\usage{
sacagrid(color = 2, cexlab = 0.7, etiq = TRUE)
}
\value{
Devuelve un data.frame con el punto central de cada una de las cuadriculas en la campaña de Porcupine. Un objeto llamado por defecto Porc.grid data.frame con cuatro columnas x,y,pt (número correlativo en cada estrato),strat (estrato del punto)
}
\description{
Crea el grid de cuadrículas de Porcupine con los puntos centrales de cada una de las cuadrículas (5x5 millas náuticas) disponibles en la campaña de Porcupine
}
\seealso{
{\link{sorteo}}
}
